+++ src-tauri/src/lib_commands/app_api/mod.rs
@@ -2,12 +2,14 @@ use super::*;
 
 mod core;
 mod navidrome;
+mod perf;
 mod remote;
 mod integration;
 mod analysis;
 
 pub(crate) use core::*;
 pub(crate) use navidrome::*;
+pub(crate) use perf::*;
 pub(crate) use remote::*;
 pub(crate) use integration::*;
 pub(crate) use analysis::*;
